feat: 移除 runtime_sync 模块并重构设备运行时协调逻辑 - 移除 runtime_sync.rs 文件及相关函数 - 在 process_control.rs 中添加设备运行时协调相关函数 - 在 infrastructure.rs 中移除 runtime_sync 模块引用 - 在 process_api.rs 中更新 dispatch_planner 的导入 - 在 session_builder.rs 中修复执行策略的类型转换 - 在 lib.rs 中更新 MainProcessCtx 的初始化 - 在 device.ts 和 domain.ts 中添加设备运行时协调相关类型 - 在 DeviceList.vue 和 ScriptEditor.vue 中重构设备表构建逻辑 - 在 TaskManagement.vue 和 TaskDevicePanel.vue 中更新设备状态显示 - 添加设备表单验证逻辑到 deviceFormValidation.ts 文件
BREAKING CHANGE: 移除了 runtime_sync 模块，相关功能已整合至 process_control.rs 中。

+++ src-tauri/crates/runtime_engine/src/domain/devices/device_conf.rs
@@ -36,7 +36,7 @@ pub struct DeviceConfig {
     #[serde(default)]
     pub transport_kind: DeviceTransportKind,
     #[serde(default = "default_startup_delay_secs")]
-    pub startup_delay_secs: u64,
+    pub startup_delay_secs: u32,
     #[serde(default)]
     #[ts(as = "Option<String>")]
     pub connect_address: Option<SocketAddrV4>,
@@ -75,9 +75,9 @@ pub struct DeviceConfig {
 #[ts(export)]
 #[serde(rename_all = "camelCase")]
 pub struct DeviceExecutionPolicy {
-    pub action_wait_ms: u64,
+    pub action_wait_ms: u32,
     pub progress_timeout_enabled: bool,
-    pub progress_timeout_ms: u64,
+    pub progress_timeout_ms: u32,
     pub timeout_action: TimeoutAction,
     pub timeout_notify_channels: Vec<TimeoutNotifyChannel>,
 }
@@ -195,7 +195,7 @@ impl Default for DeviceExecutionPolicy {
     }
 }
 
-fn default_startup_delay_secs() -> u64 {
+fn default_startup_delay_secs() -> u32 {
     15
 }
